fix(migrations): prepare phase3 revision

+++ alembic/versions/c0b3004ec4d7_add_closed_at_to_tasks.py
@@ -5,8 +5,9 @@ Revises: 26e0239a97e0
 Create Date: 2025-12-07 00:00:00.000000
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
+from sqlalchemy import inspect
 
 
 # revision identifiers, used by Alembic.
@@ -17,8 +18,14 @@ depends_on = None
 
 
 def upgrade():
-    op.add_column("tasks", sa.Column("closed_at", sa.DateTime(), nullable=True))
+    bind = op.get_bind()
+    inspector = inspect(bind)
+    if "tasks" in inspector.get_table_names():
+        op.add_column("tasks", sa.Column("closed_at", sa.DateTime(), nullable=True))
 
 
 def downgrade():
-    op.drop_column("tasks", "closed_at")
+    bind = op.get_bind()
+    inspector = inspect(bind)
+    if "tasks" in inspector.get_table_names():
+        op.drop_column("tasks", "closed_at")
